+++ backend/src/main/java/api/clue/controller/PaperController.java
@@ -2,12 +2,12 @@ package api.clue.controller;
 
 import api.clue.domain.Paper;
 import api.clue.domain.PaperSearchProvider;
+import api.clue.domain.User;
 import api.clue.service.PaperService;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.regex.Pattern;
-import org.apache.ibatis.annotations.Param;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -49,21 +49,33 @@ public class PaperController {
   }
 
   @PostMapping(value = "")
-  public void add(@RequestBody Paper paper) {
-    this.paperService.add(paper);
+  public ResponseEntity<Void> add(@AuthenticationPrincipal User user, @RequestBody Paper paper) {
+    if (user != null) {
+      this.paperService.add(paper);
+      return new ResponseEntity<>(HttpStatus.OK);
+    }
+    return new ResponseEntity<>(HttpStatus.FORBIDDEN);
   }
 
   @PatchMapping(value = "/{paperId}")
-  public void set(@PathVariable Long paperId, @RequestBody Paper paper) {
-    if (paperId != paper.getPaperId()) {
-      paper.setPaperId(paperId);
+  public ResponseEntity<Void> set(@AuthenticationPrincipal User user, @PathVariable Long paperId, @RequestBody Paper paper) {
+    if (user != null) {
+      if (!paperId.equals(paper.getPaperId())) {
+        paper.setPaperId(paperId);
+      }
+      this.paperService.set(paper);
+      return new ResponseEntity<>(HttpStatus.OK);
     }
-    this.paperService.set(paper);
+    return new ResponseEntity<>(HttpStatus.FORBIDDEN);
   }
 
   @DeleteMapping("/{paperId}")
-  public void remove(@PathVariable Long paperId) {
-    this.paperService.remove(paperId);
+  public ResponseEntity<Void> remove(@AuthenticationPrincipal User user, @PathVariable Long paperId) {
+    if (user != null) {
+      this.paperService.remove(paperId);
+      return new ResponseEntity<>(HttpStatus.OK);
+    }
+    return new ResponseEntity<>(HttpStatus.FORBIDDEN);
   }
 
 }
